whatsaap_backend/requirements.txt
fastapi
uvicorn
python-dotenv
twilio
sqlalchemy
psycopg2-binary
